// mogadishu rentals/types.ts
export interface Product {
  id: string;
  title: string;
  price: number;
  location: string;
  address?: string;
  city?: string;
  state?: string;
  image: string;
  category: string;
  photos?: string[];
  verified?: boolean;
  isNew?: boolean;
  description?: string;
  saved?: boolean;
  condition?: 'New' | 'Used';
  createdAt?: string;
  bedrooms?: number;
  bathrooms?: number;
  areaSqft?: number;
  isVisible?: boolean;
  isDeleted?: boolean;
  ownerId?: string;
}